fapimqtt/requirements.txt
fastapi >= 0.72.2
uvicorn >= 0.17.2
pydantic == 1.9.0
pymongo >= 4.1.1
paho-mqtt >= 1.6.1